Fix parsing of non UTF-8 characters in incoming invites
When a peer short invite has a non UTF-8 character, he would get a 500 error. Now the
backend just ignores the peer name field and passes an empty string to retroshare-service.

Also changed the build system to use a gradle node plugin and added the chatserver github
repository to the links page.